Building imagemagick 7.1.0-16 entrypoint

+++ .github/actions/rockylinux-build/entrypoint.sh
@@ -5,18 +5,18 @@ if [ -z "$GITHUB_WORKSPACE" ]; then
     exit 1
 fi
 
-OS=$(cat /etc/os-release | grep "PRETTY_NAME" | sed 's/PRETTY_NAME=//g' | sed 's/["]//g' | awk '{print $1}')
-
 DEST_DIR=$GITHUB_WORKSPACE/packages
 
-echo "Copying prebuilt RPMs for $IMAGEMAGICK_VERSION to $DEST_DIR"
+echo "Copying prebuilt packages for $IMAGEMAGICK_VERSION to $DEST_DIR"
 mkdir -p "$DEST_DIR"
 
-if [ "$OS" = "Ubuntu" ]; then \
+if [ -e /etc/debian_version ]; then
    cp /root/*.deb "$DEST_DIR" 
-else 
-   cp /root/rpmbuild/RPMS/x86_64/*.rpm  "$DEST_DIR"
-fi
+else
+   if [ -e /etc/redhat-release ]; then
+      cp /root/rpmbuild/RPMS/x86_64/*.rpm  "$DEST_DIR"
+   fi
+ fi
 
 echo "::set-output name=built-version::$IMAGEMAGICK_VERSION"
